06-data-infrastructure-for-gen-ai-app/pyproject.toml
[project]
name = "06-data-infrastructure-for-gen-ai-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Kan Ouivirach",email = "zkan.cs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai (>=1.14.0,<2.0.0)",
    "google-cloud-bigquery (>=3.34.0,<4.0.0)",
    "pandas-gbq (>=0.29.2,<0.30.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
